Fri, Apr 26, 2024, 12:40 PM +08:00

+++ L4/ps4/filter-less/helpers.c
@@ -11,9 +11,9 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         {
             float average = (image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed) / 3;
 
-            image[i][j].rgbtBlue = round(average);
-            image[i][j].rgbtGreen = round(average);
-            image[i][j].rgbtRed = round(average);
+            image[i][j].rgbtBlue = (int)round(average);
+            image[i][j].rgbtGreen = (int)round(average);
+            image[i][j].rgbtRed = (int)round(average);
         }
     }
     return;
@@ -33,17 +33,17 @@ void sepia(int height, int width, RGBTRIPLE image[height][width])
             if (tmpB > 255)
                 image[i][j].rgbtBlue = 255;
             else
-                image[i][j].rgbtBlue = round(tmpB);
+                image[i][j].rgbtBlue = (int)round(tmpB);
 
             if (tmpG > 255)
                 image[i][j].rgbtGreen = 255;
             else
-                image[i][j].rgbtGreen = round(tmpG);
+                image[i][j].rgbtGreen = (int)round(tmpG);
 
             if (tmpR > 255)
                 image[i][j].rgbtRed = 255;
             else
-                image[i][j].rgbtRed = round(tmpR);
+                image[i][j].rgbtRed = (int)round(tmpR);
         }
     }
     return;
@@ -117,8 +117,8 @@ RGBTRIPLE box_blur(int h, int w, int height, int width, RGBTRIPLE copy[height][w
         }
     }
 
-    new_pixel.rgbtRed = round(sum[0] / num);
-    new_pixel.rgbtGreen = round(sum[1] / num);
-    new_pixel.rgbtBlue = round(sum[2] / num);
+    new_pixel.rgbtRed = (int)round(sum[0] / num);
+    new_pixel.rgbtGreen = (int)round(sum[1] / num);
+    new_pixel.rgbtBlue = (int)round(sum[2] / num);
     return new_pixel;
 }
